Fix yyset_lineno() Flex compatibility wrappers

+++ include/reflex/flexlexer.h
@@ -365,7 +365,7 @@ int yywrap(void);
 #define yyget_lineno(s)         static_cast<int>(static_cast<FlexLexer*>(s)->lineno())
 
 /// Flex-compatible macro: set the lineno accessor of a reentrant scanner.
-#define yyset_lineno(n,s)       static_cast<int>(static_cast<FlexLexer*>(s)->lineno(n))
+#define yyset_lineno(n,s)       static_cast<FlexLexer*>(s)->lineno(n)
 
 /// Flex-compatible macro: the reflex::Input in accessor of a reentrant scanner.
 #define yyget_in(s)             static_cast<FlexLexer*>(s)->in()
